Fix upscale error handling
Upscale process uses universal line endings which results in a str result from the stdout stream instead of bytes.  We need to not decode the output.  Also the output is being processed by the monitoring loop so just build a list of output for passing along in the exception to aid debugging.