Hacky Fix for weird aspect ratio encodes
As is typical of files encoded by handbrake, some video files may have weird display aspect ratios and sample aspect ratios due to attempts to crop the video of black bars along the edges.  If we detect a display aspect ratio mismatch, then fudge the scaling of the source video so that it aligns with a multiple of the display aspect ratio.  If padding is enabled, this shouldn't be necessary.